requirements.txt
asgiref==3.6.0
Django==4.2.1
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.4.0
xlrd==2.0.1
XlsxWriter==3.1.2
xlwt==1.3.0
langdetect==1.0.9
redis==5.0.1
gunicorn==21.2.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
PyYAML==6.0.1
tablib==3.5.0
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
django-import-export==3.2.0
et-xmlfile==1.1.0
backports.zoneinfo==0.2.1;python_version<"3.9"
psycopg2-binary==2.9.9